api/requirements.txt
pip>=9
setuptools>=26
twine
wheel>=0.29

streamlit
fastapi
google-cloud-vision
python-multipart
python-dotenv
requests
uvicorn